+++ packages/kitchen/CHANGELOG.md
@@ -1,5 +1,80 @@
 # @tonightpass/kitchen
 
+## 1.3.0
+
+### Minor Changes
+
+- 0a274f6: Native | Add `Container`
+
+## 1.2.4
+
+### Patch Changes
+
+- f85297c: Native | Fix type light text color on `Button`
+
+## 1.2.3
+
+### Patch Changes
+
+- 400bd8d: Fix scale hoc types
+
+## 1.2.2
+
+### Patch Changes
+
+- 6b6aa02: Native | Fix `withScale` extension from `.ts` to `.tsx`
+- 6b6aa02: Fix `my` prop from `withScale`
+- 8544095: Fix global types errors
+
+## 1.2.1
+
+### Patch Changes
+
+- cb33b26: Native | Fix missing react import into `Button`
+
+## 1.2.0
+
+### Minor Changes
+
+- 83f386f: Native | Add `Button`
+
+## 1.1.7
+
+### Patch Changes
+
+- 0e0b8e0: Native | Fix paths in package.json
+
+## 1.1.6
+
+### Patch Changes
+
+- f4ab4d6: Add types on each entrypoints
+
+## 1.1.5
+
+### Patch Changes
+
+- c9c82b1: Native | Fix `Text` line-height
+
+## 1.1.4
+
+### Patch Changes
+
+- ecb2128: Native | Fix `Text` props
+
+## 1.1.3
+
+### Patch Changes
+
+- dd594cf: Native | Fix `Text` transform props
+
+## 1.1.2
+
+### Patch Changes
+
+- 4e096e9: Add `native` theme detection
+- f317293: Update `native` themes
+
 ## 1.1.1
 
 ### Patch Changes